agora_mobile/requirements.txt
kivy==2.2.1
kivymd==1.1.1
requests==2.31.0
python-jose==3.3.0
bcrypt==4.1.1
sqlalchemy==2.0.23
aiohttp==3.9.1
python-dotenv==1.0.0
pillow==10.1.0
openai==1.6.1
anthropic==0.7.0
langchain==0.0.340
langchain-openai==0.0.2
buildozer==1.5.0 